add batch.yml config file, parse in main.cpp and perform specified actions - update kernels to be int** not double**

+++ src/filter/median_filter.cpp
@@ -2,7 +2,7 @@
 
 using namespace cv;
 
-void median_filter(Mat *src, double **kernel, int kernel_size) {
+void median_filter(Mat *src, int **kernel, int kernel_size) {
     Mat dst = Mat::ones(src->rows, src->cols, CV_8U);
 
     uint8_t mask_size = 3; // TODO: user should input
